makes the script rerunnable 'idempotent'ish

+++ install.sh
@@ -10,12 +10,22 @@ ln -s -f ~/proj/env/.gvimrc ~/.gvimrc
 ln -s -f ~/proj/env/.gitconfig ~/.gitconfig
 
 # map capslock to esc 
-echo "setxkbmap -option caps:escape" > ~/.xsessionrc
+if ! grep --quiet "setxkbmap -option caps:escape" ~/.xsessionrc; then
+    echo "setxkbmap -option caps:escape" >> ~/.xsessionrc
+fi
 
 # setup vundle from cloning in git
-mkdir -p ~/.vim/bundle
-git clone https://github.com/VundleVim/Vundle.vim.git ~/.vim/bundle/Vundle.vim
+if [ -d ~/.vim/bundle ]; then
+    pushd .
+    cd ~/.vim/bundle/Vundle.vim
+    git pull
+    popd
+else
+    mkdir -p ~/.vim/bundle
+    git clone https://github.com/VundleVim/Vundle.vim.git ~/.vim/bundle/Vundle.vim
+fi
 vim +PluginInstall! +qall
+vim +PluginClean! +qall
 
 # setup bash profile
 ln -s -f ~/proj/env/.bash_aliases ~/.bash_aliases
@@ -23,10 +33,12 @@ ln -s -f ~/proj/env/.bash_profile ~/.bash_profile
 
 # install chrome
 # from http://www.tecmint.com/install-google-chrome-in-debian-ubuntu-linux-mint/
-wget -q -O - https://dl-ssl.google.com/linux/linux_signing_key.pub | sudo apt-key add -
-sudo sh -c 'echo "deb http://dl.google.com/linux/chrome/deb/ stable main" >> /etc/apt/sources.list.d/google.list'
-sudo apt-get update
-sudo apt-get install google-chrome-stable -y
+if ! dpkg -s google-chrome-stable &> /dev/null; then
+    wget -q -O - https://dl-ssl.google.com/linux/linux_signing_key.pub | sudo apt-key add -
+    sudo sh -c 'echo "deb http://dl.google.com/linux/chrome/deb/ stable main" >> /etc/apt/sources.list.d/google.list'
+    sudo apt-get update
+    sudo apt-get install google-chrome-stable -y
+fi
 
 # install consolas font
 mkdir -p ~/.fonts
@@ -37,20 +49,24 @@ ln -s -f ~/proj/env/CONSOLAZ.TTF ~/.fonts/CONSOLAAZ.TTF
 fc-cache -f -v
 
 # install dropbox
-pushd .
-cd ~ && wget -O - "https://www.dropbox.com/download?plat=lnx.x86_64" | tar xzf -
-~/.dropbox-dist/dropboxd
-popd
+if [ ! -d ~/.dropbox-dist ]; then
+    pushd .
+    cd ~ && wget -O - "https://www.dropbox.com/download?plat=lnx.x86_64" | tar xzf -
+    ~/.dropbox-dist/dropboxd
+    popd
+fi
 
 # password safe
-pushd .
-cd ~
-wget "http://freefr.dl.sourceforge.net/project/passwordsafe/Linux-BETA/0.94/passwordsafe-ubuntu-0.94BETA.amd64.deb"
-sudo dpkg -i passwordsafe-ubuntu-0.94BETA.amd64.deb -y
-sudo apt-get -f install -y
-sudo dpkg -i passwordsafe-ubuntu-0.94BETA.amd64.deb -y
-rm passwordsafe-ubuntu-0.94BETA.amd64.deb
-popd
+if [ ! -d ~/.dropbox-dist ]; then
+    pushd .
+    cd ~
+    wget "http://freefr.dl.sourceforge.net/project/passwordsafe/Linux-BETA/0.94/passwordsafe-ubuntu-0.94BETA.amd64.deb"
+    sudo dpkg -i passwordsafe-ubuntu-0.94BETA.amd64.deb -y
+    sudo apt-get -f install -y
+    sudo dpkg -i passwordsafe-ubuntu-0.94BETA.amd64.deb -y
+    rm passwordsafe-ubuntu-0.94BETA.amd64.deb
+    popd
+fi
 
 # install nodejs
 sudo apt-get install nodejs -y
@@ -71,9 +87,15 @@ function addKeyboardApp()
     cp $path.new $path
 }
 
-addKeyboardApp "<Primary><Alt>g" "gvim"
-addKeyboardApp "<Primary><Alt>t" "xfce4-terminal --maximize --fullscreen --hide-menubar"
-addKeyboardApp "<Primary><Alt>b" "/usr/bin/google-chrome-stable"
+if ! grep --quiet "gvim" ~/.config/xfce4/xfconf/xfce-perchannel-xml/xfce4-keyboard-shortcuts.xml; then
+    addKeyboardApp "<Primary><Alt>g" "gvim"
+fi
+if ! grep --quiet "xfce4-terminal" ~/.config/xfce4/xfconf/xfce-perchannel-xml/xfce4-keyboard-shortcuts.xml; then
+    addKeyboardApp "<Primary><Alt>t" "xfce4-terminal --maximize --fullscreen --hide-menubar"
+fi
+if ! grep --quiet "google-chrome-stable" ~/.config/xfce4/xfconf/xfce-perchannel-xml/xfce4-keyboard-shortcuts.xml; then
+    addKeyboardApp "<Primary><Alt>b" "/usr/bin/google-chrome-stable"
+fi
 
 # cleanup
 sudo apt-get autoremove -y
@@ -82,18 +104,11 @@ sudo apt-get autoremove -y
 
 # setup password safe and dropbox to get my passwords
 
-# setup launcher shortcuts:
-# ctrl-alt-t = xfce4-terminal --maximize --fullscreen
-# ctrl-alt-g = gvim -f
-# ctrl-alt-b = /usr/bin/google-chrome-stable %U
-
 # setup bash prompt to include git info
 # setup a git credential manager to remember git password
 
 # set terminal config to hide toolbar and chrome
 # set terminal config to load maximized
-# set a terminal shortcut key
 
-# link gvimrc
 # setup xmllint for prertyxml in vim
